Refactor terminal type into stand-alone function. Move non-compiler code back to blocklyc.js Remove the 'terminalNeeded' parameter in the compiler callback.
