+++ src/modules/blocklyc.js
@@ -32,6 +32,7 @@ import {getProjectInitialState} from './project';
 import {logConsoleMessage, getURLParameter, utils} from './utility';
 import {sanitizeFilename} from './utility';
 import {cloudCompile} from './compiler';
+import {serialConsole} from './serial_console';
 
 /**
  * TODO: Identify the purpose of this variable
@@ -173,145 +174,6 @@ const graph_data = {
   ],
 };
 
-// /**
-//  * Submit a project's source code to the cloud compiler
-//  *
-//  * @param {string} text
-//  * @param {string} action
-//  * @param {function} successHandler Define a callback to be executed upon
-//  *  sucessful compilation
-//  */
-// function cloudCompile(text, action, successHandler) {
-//   const codePropC = getSourceEditor();
-//   const project = getProjectInitialState();
-//   // if PropC is in edit mode, get it from the editor, otherwise
-//   // render it from the blocks.
-//   let propcCode = '';
-//
-//   if (codePropC.getReadOnly()) {
-//     propcCode = prettyCode(
-//         Blockly.propc.workspaceToCode(Blockly.mainWorkspace));
-//   } else {
-//     propcCode = codePropC.getValue();
-//   }
-//
-//   if (propcCode.indexOf('EMPTY_PROJECT') > -1) {
-//     utils.showMessage(Blockly.Msg.DIALOG_EMPTY_PROJECT,
-//         Blockly.Msg.DIALOG_CANNOT_COMPILE_EMPTY_PROJECT);
-//   } else {
-//     $('#compile-dialog-title').text(text);
-//     $('#compile-console').val('Compile... ');
-//     $('#compile-dialog').modal('show');
-//
-//     let terminalNeeded = null;
-//
-//     // TODO: propc editor needs UI for settings for terminal and graphing
-//     if (project.boardType.name !== 'propcfile') {
-//       const consoleBlockList = [
-//         'console_print', 'console_print_variables', 'console_print_multiple',
-//         'console_scan_text', 'console_scan_number', 'console_newline',
-//         'console_clear', 'console_move_to_position', 'oled_font_loader',
-//         'activitybot_display_calibration', 'scribbler_serial_send_text',
-//         'scribbler_serial_send_char', 'scribbler_serial_send_decimal',
-//         'scribbler_serial_send_decimal', 'scribbler_serial_send_ctrl',
-//         'scribbler_serial_cursor_xy',
-//       ];
-//
-//       let consoleBlockCount = 0;
-//       for (let i = 0; i < consoleBlockList.length; i++) {
-//         consoleBlockCount += Blockly.getMainWorkspace()
-//             .getBlocksByType(consoleBlockList[i], false).length;
-//       }
-//
-//       if (consoleBlockCount > 0) {
-//         terminalNeeded = 'term';
-//       } else if (Blockly.getMainWorkspace()
-//           .getBlocksByType('graph_settings', false).length > 0) {
-//         terminalNeeded = 'graph';
-//       }
-//     }
-//
-//     // ------------------------------------------------------------------------
-//     // Contact the container running cloud compiler. If the browser is
-//     // connected via https, direct the compile request to the same port and let
-//     // the load balancer direct the request to the compiler.
-//     // When operating from localhost, expect to find the compiler container on
-//     // localhost as well. There is no override for this at the moment.
-//     // ------------------------------------------------------------------------
-//     let postUrl = `https://${window.location.hostname}:443/single/prop-c/${action}`;
-//     if (window.location.protocol === 'http:') {
-//       postUrl = `http://${window.location.hostname}:5001/single/prop-c/${action}`;
-//     }
-//
-//     // Post the code to the compiler API and await the results
-//     logConsoleMessage(`Requesting compiler service`);
-//     $.ajax({
-//       'method': 'POST',
-//       'url': postUrl,
-//       'data': {'code': propcCode},
-//     }).done(function(data) {
-//       logConsoleMessage(`Receiving compiler service results`);
-//       // The compiler will return one of three payloads:
-//       // Compile-only
-//       // data = {
-//       //     "success": success,
-//       //     "compiler-output": out,
-//       //     "compiler-error": err.decode()
-//       // }
-//       //
-//       // Load to RAM/EEPROM
-//       // data = {
-//       //     "success": success,
-//       //     "compiler-output": out,
-//       //     "compiler-error": err.decode()
-//       //     "binary": base64binary.decode('utf-8')
-//       //     "extension": = extension
-//       // }
-//       //
-//       // General error message
-//       // data = {
-//       //    "success": False,
-//       //    "message": "unknown-action",
-//       //    "data": action
-//       // }
-//       // {success: true, compiler-output: "Succeeded.", compiler-error: ""}
-//
-//       // Check for an error response from the compiler
-//       if (!data || data['compiler-error'] != '') {
-//         // Get message as a string, or blank if undefined
-//         const message = (typeof data['compiler-error'] === 'string') ?
-//             data['compiler-error'] : '';
-//         appendCompileConsoleMessage(
-//             data['compiler-output'] + data['compiler-error'] + message);
-//       } else {
-//         const loadWaitMsg = (action !== 'compile') ? '\nDownload...' : '';
-//         appendCompileConsoleMessage(
-//             data['compiler-output'] + data['compiler-error'] + loadWaitMsg);
-//
-//         if (data.success && successHandler) {
-//           successHandler(data, terminalNeeded);
-//         }
-//         compileConsoleScrollToBottom();
-//       }
-//     }).fail(function(data) {
-//       // Something unexpected has happened while calling the compile service
-//       if (data) {
-//         logConsoleMessage(`Compiler service request failed: ${data.state()}`);
-//
-//         const state = data.state();
-//         let message = 'Unable to compile the project.\n';
-//         if (state === 'rejected') {
-//           message += '\nThe compiler service is temporarily unavailable or';
-//           message += ' unreachable.\nPlease try again in a few moments.';
-//         } else {
-//           message += 'Error "' + data.status + '" has been detected.';
-//         }
-//         appendCompileConsoleMessage(message);
-//       }
-//     });
-//   }
-// }
-
 /**
  * This is the onClick event handler for the compile toolbar button
  */
@@ -320,7 +182,8 @@ export function compile() {
 }
 
 /**
- * Begins loading process
+ * The onClick event handler for the Compile to RAM and Compile to EEPROM
+ * UI toolbar buttons of the same names.
  *
  * @param {string} modalMessage message shown at the top of the
  *  compile/load modal.
@@ -335,8 +198,9 @@ export function compile() {
  */
 export function loadInto(modalMessage, compileCommand, loadOption, loadAction) {
   if (clientService.portsAvailable) {
-    cloudCompile(modalMessage, compileCommand, function(data, terminalNeeded) {
+    cloudCompile(modalMessage, compileCommand, function(data) {
       logConsoleMessage(`Processing cloud compiler callback`);
+      const terminalNeeded = isTerminalWindowRequired();
 
       if (clientService.type === serviceConnectionTypes.WS) {
         logConsoleMessage(`(LOAI) Device loaded via websocket`);
@@ -410,9 +274,8 @@ export function loadInto(modalMessage, compileCommand, loadOption, loadAction) {
             'extension': data.extension,
             'comport': getComPort(),
           }, function(loadData) {
-            $('#compile-console')
-                .val($('#compile-console')
-                    .val() + loadData.message);
+            $('#compile-console').val($('#compile-console').val() +
+                loadData.message);
 
             // Scroll automatically to the bottom after new data is added
             document.getElementById('compile-console').scrollTop =
@@ -436,6 +299,46 @@ export function loadInto(modalMessage, compileCommand, loadOption, loadAction) {
         Blockly.Msg.DIALOG_DEVICE_COMM_ERROR_TEXT);
   }
 }
+
+
+/**
+ * Evaluate the project to determine if a terminal or graph window is required
+ * when the project is run on the device
+ *
+ * @return {string}
+ */
+const isTerminalWindowRequired = () => {
+  const project = getProjectInitialState();
+  let terminalNeeded = '';
+
+  // TODO: propc editor needs UI for settings for terminal and graphing
+  if (project.boardType.name !== 'propcfile') {
+    const consoleBlockList = [
+      'console_print', 'console_print_variables', 'console_print_multiple',
+      'console_scan_text', 'console_scan_number', 'console_newline',
+      'console_clear', 'console_move_to_position', 'oled_font_loader',
+      'activitybot_display_calibration', 'scribbler_serial_send_text',
+      'scribbler_serial_send_char', 'scribbler_serial_send_decimal',
+      'scribbler_serial_send_decimal', 'scribbler_serial_send_ctrl',
+      'scribbler_serial_cursor_xy',
+    ];
+
+    let consoleBlockCount = 0;
+    for (let i = 0; i < consoleBlockList.length; i++) {
+      consoleBlockCount += Blockly.getMainWorkspace()
+          .getBlocksByType(consoleBlockList[i], false).length;
+    }
+
+    if (consoleBlockCount > 0) {
+      terminalNeeded = 'term';
+    } else if (Blockly.getMainWorkspace()
+        .getBlocksByType('graph_settings', false).length > 0) {
+      terminalNeeded = 'graph';
+    }
+  }
+  return terminalNeeded;
+};
+
 //
 // /**
 //  * Serial console support
@@ -1262,3 +1165,4 @@ export function compileConsoleScrollToBottom() {
   const compileConsoleObj = document.getElementById('compile-console');
   compileConsoleObj.scrollTop = compileConsoleObj.scrollHeight;
 }
+
